api/requirements.txt
catkin==0.7.29
catkin-pkg==0.5.2
click==7.1.2
docutils==0.18.1
Flask==1.1.4
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
message-filters==1.14.13
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.18.0
PyYAML==5.4.1
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospkg==1.4.0
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
sensor-msgs==1.12.8
six==1.16.0
topic-tools==1.14.13
typing==3.10.0.0
Werkzeug==1.0.1
